requirements.txt
# Django and REST Framework (Python 3.7 compatible)
Django>=3.2,<4.0
djangorestframework>=3.12.0,<3.15
djangorestframework-simplejwt>=4.8.0,<5.0

# Database
PyMySQL>=1.0.0

# Redis and Celery
redis>=4.0.0,<5.0
celery>=5.2.0,<5.3
django-redis>=5.2.0,<5.3

# API Documentation
drf-spectacular>=0.24.0,<0.26

# Rate Limiting
django-ratelimit>=3.0.0,<4.0

# CORS
django-cors-headers>=3.13.0,<4.0

# Testing
pytest>=7.0.0,<8.0
pytest-django>=4.5.0,<4.6
hypothesis>=6.50.0,<6.80

# Utilities
python-decouple>=3.6
